More fixes for stupid stuff, also temporarily loosen interface to avoid BC breaks
