+++ src/Discussion/Search/Gambit/FulltextGambit.php
@@ -9,12 +9,10 @@
 
 namespace Flarum\Discussion\Search\Gambit;
 
-use Flarum\Discussion\Search\DiscussionSearch;
 use Flarum\Post\Post;
 use Flarum\Search\GambitInterface;
 use Flarum\Search\SearchState;
 use Illuminate\Database\Query\Expression;
-use LogicException;
 
 class FulltextGambit implements GambitInterface
 {
@@ -23,10 +21,6 @@ class FulltextGambit implements GambitInterface
      */
     public function apply(SearchState $search, $bit)
     {
-        if (! $search instanceof DiscussionSearch) {
-            throw new LogicException('This gambit can only be applied on a DiscussionSearch');
-        }
-
         // Replace all non-word characters with spaces.
         // We do this to prevent MySQL fulltext search boolean mode from taking
         // effect: https://dev.mysql.com/doc/refman/5.7/en/fulltext-boolean.html
