+++ src/Search/AbstractSearch.php
@@ -1,0 +1,45 @@
+<?php
+
+/*
+ * This file is part of Flarum.
+ *
+ * For detailed copyright and license information, please view the
+ * LICENSE file that was distributed with this source code.
+ */
+
+namespace Flarum\Search;
+
+use Flarum\Filter\FilterState;
+
+/**
+ * @deprecated, use SearchState instead.
+ * These methods should be transferred over to SearchState in beta 17.
+ */
+class AbstractSearch extends FilterState
+{
+    /**
+     * @var GambitInterface[]
+     */
+    protected $activeGambits = [];
+
+    /**
+     * Get a list of the gambits that are active in this search.
+     *
+     * @return GambitInterface[]
+     */
+    public function getActiveGambits()
+    {
+        return $this->activeGambits;
+    }
+
+    /**
+     * Add a gambit as being active in this search.
+     *
+     * @param GambitInterface $gambit
+     * @return void
+     */
+    public function addActiveGambit(GambitInterface $gambit)
+    {
+        $this->activeGambits[] = $gambit;
+    }
+}
