+++ src/Search/SearchState.php
@@ -9,33 +9,4 @@
 
 namespace Flarum\Search;
 
-use Flarum\Filter\FilterState;
-
-abstract class SearchState extends FilterState
-{
-    /**
-     * @var GambitInterface[]
-     */
-    protected $activeGambits = [];
-
-    /**
-     * Get a list of the gambits that are active in this search.
-     *
-     * @return GambitInterface[]
-     */
-    public function getActiveGambits()
-    {
-        return $this->activeGambits;
-    }
-
-    /**
-     * Add a gambit as being active in this search.
-     *
-     * @param GambitInterface $gambit
-     * @return void
-     */
-    public function addActiveGambit(GambitInterface $gambit)
-    {
-        $this->activeGambits[] = $gambit;
-    }
-}
+class SearchState extends AbstractSearch {}
